fix: don't run opensearch-cli in a child process
The agent CLI starts opensearch-cli in a child process. This has some
nasty drawbacks:

- any signal passed to the agent is not forwarded to opensearch-cli,
  which forces one to kill the process group to actually kill the PA
- the exit code of the agent is always 0, and the actual exit code of
  opensearch-cli is lost

This commit execs the opensearch-cli command instead of running it in a
subprocess, so that any signal is forwarded to the callee, and any exit
code is forwarded to the caller.

See https://github.com/opensearch-project/opensearch-build/issues/1487

Signed-off-by: Gilbert Gilb's <gilbsgilbert@gmail.com>